pyproject.toml
[project]
name = "apcragg-blog"
version = "0.1.0"
description = "My blog"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "markdown>=3.10",
    "numpy>=2.3.5",
    "pygments>=2.19.2",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "watchdog>=6.0.0",
]
